LayoutTests/fast/mediastream/enumerateDevices-camera-denied-expected.txt


PASS enumerateDevices() filters out cameras if camera permission is prompt
PASS enumerateDevices() filters out cameras if camera permission is denied
PASS enumerateDevices() does not filter out cameras if camera permission is granted
